renamed layer to Layers moved parent class outside of the folder Layers
