+++ src/architecture/Layers/HopfieldLayer.ts
